requirements.txt
numpy 
pandas 
matplotlib.pyplot 
seaborn 
seaborn
matplotlib.pyplot 